feat(behavior): add new event based gateway behavior
Handles the scenario where a user wants to connect an event-based
gateway to another event-based gateway target with existing connections.
The existing connections are removed before adding the new one.

This is because target elements in an event gateway configuration
must not have any additional incoming sequence flows other than
that from the event gateway.

The implementation handles multiple existing connections because this
could be modeled in an external tool which doesn't prevent multiple
incoming connections to target events of an event based gateway target.